requirements.txt
Flask
googletrans==4.0.0-rc1
python-dotenv